Major update v0.2
Changed the index page to be the login page as that's where we want the flow of people to go through first and foremost. Created a log in to bypass the login screen for testing: Username is testuser@gmail.com and the password is testpass. Once you're in the website the buttons for home, my bets, and live in bets, all send you to the corresponding page with a indicator that follows your specific page that you are in. also altered the top navigation to fit better with the black navigation bar. The log in button also sends you back to the log in page, (works as like a log out button for the time being). In turn with this update I had to revert my changes on the sidebar and we can decide how to change and alter it moving forward. also just as a slight change I altered the notifications icon to be a gear as it'll be a settings button in the future. I made a CSS file to manage the top bars as they should be identical and consistent across almost every page (not including the login etc). the only thing that I particularly had ideas to change in the future is like nitpicks but it is to fix the shifting that happens when going to / from the home page from another page. and to make the animation for page switching in the navigation smoother rather than sharp as it is currently. Yes the live in bets page also had to be scrapped to continue this update but we can work to implement many changes and additions to it easily as we did before.