file directory update and home page tweaks
I made file directories to make this hopefully be a smoother experience to update in the future. but also that means i had to tweak the home page a little bit and tweak the sidebar. but from now on it should be lighter and simpler to update and add and create.